Port this container to certbot.
The old letsencrypt client has been replaced by the way more powerful
client called certbot. That draws this container very minimalistic, as
priorily we needed to check, which certificates need to be updated, now
certbot does that for us. All we need to do is to start cron with a
crontab to call `certbot renew` regularily and to provide a volume with
the old certificates available. certbot will then do the rest. We don't
even need to configure it, as it will for each certificate use the
configuration that was used, back when the certificate was issued the
first time.